models/c_models/data_loader/pyproject.toml
[project]
name = "data-loader"
version = "0.1.0"
description = "Data loader for Christians models."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "db-dtypes>=1.4.3",
    "google-cloud-bigquery",
    "pandas",
    "python-dotenv",
]
